refactor : moveStrategy interface 생성
