+++ src/main/java/racingcar/repository/moveStrategy.java
@@ -1,0 +1,5 @@
+package racingcar.repository;
+
+public interface moveStrategy {
+    public void move();
+}
